+++ controller/Changelog.md
@@ -3,6 +3,7 @@
 ## Changed
 
 - driver: Upgrade to 2.3.0 for recent versions of Senso Flex
+- os: Update nixpkgs channel to 22.11
 
 ## Fixed